crypto bot payment handling

+++ src/presentation/web_api/schema/payment.py
@@ -1,5 +1,7 @@
+from decimal import Decimal
+
 from pydantic import BaseModel
 
 
 class CreateCryptoPayInvoiceSchema(BaseModel):
-    amount: int
+    amount: Decimal
